+++ src/signatures.rs
@@ -115,27 +115,6 @@ signatures! {
     format = PgpMessage
     value = b"-----BEGIN PGP MESSAGE-----"
 
-    format = StereolithographyBinary
-    value = b"3D Systems, Binary STL file"
-    value = b"Materialise Coloured STL"
-    value = b"Exported from Blender", b"\x00" offset = 79
-    value = b"Rhinoceros Binary STL"
-    value = b"Uranium STLWriter"
-    value = b"STL binary file"
-    value = b"binary stl file"
-    value = b"Stratasys stl"
-    value = b"SketchUp STL"
-    value = b"STL File"
-    value = b"STLB ATF"
-    value = b"STLEXP"
-    value = b".stl" offset = 2, b"\x00" offset = 79
-    value = b".stl" offset = 3, b"\x00" offset = 79
-    value = b".stl" offset = 4, b"\x00" offset = 79
-    value = b".stl" offset = 5, b"\x00" offset = 79
-    value = b".stl" offset = 6, b"\x00" offset = 79
-    value = b".stl" offset = 7, b"\x00" offset = 79
-    value = b".stl" offset = 8, b"\x00" offset = 79
-
     // 26 bytes
     format = TrainingCenterXml
     value = b"\xEF\xBB\xBF<TrainingCenterDatabase"